Refactor Cypress commands
