+++ pages/index.tsx
@@ -3,7 +3,6 @@ import Button from '@mui/material/Button';
 import Container from '@mui/material/Container';
 import Typography from '@mui/material/Typography';
 import { useRouter } from 'next/router';
-import React from 'react';
 import Copyright from '../src/Components/Copyright';
 import { Settings } from '../src/Components/Settings';
 
@@ -17,17 +16,19 @@ const Index: React.FC = () => {
   return (
     <Container maxWidth="sm">
       <Box my={4} sx={{ textAlign: 'center' }}>
-        <Box my={4}>
-          <Typography variant="h4" component="h1" gutterBottom>
-            Japanese Flashcards
-          </Typography>
-        </Box>
-        <Settings />
-        <Box my={4}>
-          <Button variant="outlined" onClick={handleStartGame}>
-            Play
-          </Button>
-        </Box>
+        <main aria-labelledby="title">
+          <Box my={4}>
+            <Typography id="title" variant="h4" component="h1" gutterBottom>
+              Japanese Flashcards
+            </Typography>
+          </Box>
+          <Settings />
+          <Box my={4}>
+            <Button variant="outlined" onClick={handleStartGame}>
+              Play
+            </Button>
+          </Box>
+        </main>
         <Copyright />
       </Box>
     </Container>
